+++ ant_upload_checker/output.py
@@ -6,7 +6,7 @@ import logging
 def write_film_list_to_csv(output_df):
     if not output_df.empty:
         output_file_path = Path(OUTPUT_FOLDER).joinpath("Film list.csv")
-        logging.info(f"Writing list of films to {output_file_path} ...")
+        logging.info("\nWriting list of films to %s...", output_file_path)
         output_df.to_csv(output_file_path, index=False, encoding="utf-8-sig")
     else:
-        logging.warning("The film list is empty, no file is being created.")
+        logging.warning("\nThe film list is empty, no file is being created.")